Redesign profile page
